Strip down Default Transport to improve test coverage of AbstractTransport
As per MANGO-2104
